Add custom hook errorHandler
