+++ src/hoc/errorHandler.js
@@ -1,53 +1,18 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 import Modal from '../components/layout/modal/modal';
 import Aux from './aux';
+import useErrorHandler from '../hooks/errorHandler';
 
 const errorHandler = (WrappedComponent, axios) => {
   return props => {
-    const [error, setError] = useState(null);
-
-    const reqInterceptors = axios.interceptors.request.use(request => {
-      setError(null);
-      return request;
-    });
-
-    const resInterceptors = axios.interceptors.response.use(response =>
-      response, resError => {
-        setError(resError);
-    });
-
-
-    useEffect(() => {
-      return () => {
-        axios.interceptors.request.eject(reqInterceptors);
-        axios.interceptors.response.eject(resInterceptors);
-      };
-    }, [reqInterceptors, resInterceptors])
-
-    const errorConfirmedHandler = () => {
-      setError(null);
-    };
-
-    const errorMessage = () => {
-      let errorMessage = '';
-
-      if (error.response.data.error.message) {
-        errorMessage = error.response.data.error.message;
-      } else if (error.response.data.error) {
-        errorMessage = error.response.data.error;
-      } else {
-        errorMessage = error.message;
-      }
-
-      return errorMessage;
-    };
+    const [error, errorMessage, clearError] = useErrorHandler(axios);
 
     return (
       <Aux>
         <Modal
           show={!!error}
-          modalClosed={errorConfirmedHandler}
+          modalClosed={clearError}
         >
           {error && errorMessage()}
         </Modal>
